fix(TCOMP-970): move resolveParameters() method to common super class
