+++ main/plugins/org.talend.sdk.component.studio-integration/src/main/java/org/talend/sdk/component/studio/model/parameter/resolver/AbstractParameterResolver.java
@@ -1,18 +1,22 @@
 package org.talend.sdk.component.studio.model.parameter.resolver;
 
-import java.util.Collection;
-import java.util.Collections;
-import java.util.List;
-import java.util.Map;
+import java.beans.PropertyChangeListener;
+import java.util.*;
 import java.util.stream.Collectors;
 
 import org.talend.core.model.process.IElementParameter;
+import org.talend.designer.core.model.components.ElementParameter;
 import org.talend.sdk.component.server.front.model.ActionReference;
+import org.talend.sdk.component.server.front.model.SimplePropertyDefinition;
+import org.talend.sdk.component.studio.model.action.Action;
+import org.talend.sdk.component.studio.model.action.ActionParameter;
 import org.talend.sdk.component.studio.model.parameter.PropertyDefinitionDecorator;
 import org.talend.sdk.component.studio.model.parameter.PropertyNode;
 import org.talend.sdk.component.studio.model.parameter.PropertyTreeCreator;
 import org.talend.sdk.component.studio.model.parameter.TaCoKitElementParameter;
 import org.talend.sdk.component.studio.model.parameter.WidgetTypeMapper;
+
+import static java.util.Comparator.comparing;
 
 /**
  * Common super class for ParameterResolvers. It contains common state and functionality
@@ -26,14 +30,64 @@ abstract class AbstractParameterResolver implements ParameterResolver {
     /**
      * PropertyNode, which represents Configuration class Option annotated with action annotation
      */
+    private final Action action;
+
     protected final PropertyNode actionOwner;
     
     protected final ActionReference actionRef;
+
+    private final PropertyChangeListener listener;
+
+    private final ElementParameter redrawParameter;
+
+    AbstractParameterResolver(final Action action, final PropertyNode actionOwner, final ActionReference actionRef, final PropertyChangeListener listener) {
+        this(action, actionOwner, actionRef, listener, null);
+    }
     
-    AbstractParameterResolver(final PropertyNode actionOwner, final ActionReference actionRef) {
+    AbstractParameterResolver(final Action action, final PropertyNode actionOwner, final ActionReference actionRef, final PropertyChangeListener listener,
+                              final ElementParameter redrawParameter) {
+        this.action = action;
         this.actionOwner = actionOwner;
         this.actionRef = actionRef;
+        this.listener = listener;
+        this.redrawParameter = redrawParameter;
     }
+
+    /**
+     * Finds ElementParameters needed for action call by their relative path.
+     * Registers PropertyChangeListener to each ElementParameter needed for action call
+     * Creates ActionParameter for each ElementParameter
+     *
+     * @param settings all "leaf" Component options
+     */
+    public void resolveParameters(final Map<String, IElementParameter> settings) {
+        final Iterator<PropertyDefinitionDecorator> expectedParameters = PropertyDefinitionDecorator.wrap(actionRef.getProperties())
+                .stream()
+                .filter(p -> p.getParameter().isRoot())
+                .sorted(comparing(p -> p.getParameter().getIndex()))
+                .iterator();
+        final List<String> relativePaths = getRelativePaths();
+
+        relativePaths.forEach(relativePath -> {
+            if (expectedParameters.hasNext()) {
+                final String absolutePath = pathResolver.resolvePath(getOwnerPath(), relativePath);
+                final List<TaCoKitElementParameter> parameters = resolveParameters(absolutePath, settings);
+                final SimplePropertyDefinition parameterRoot = expectedParameters.next();
+                parameters.forEach(parameter -> {
+                    parameter.registerListener("value", listener);
+                    if (redrawParameter != null) {
+                        parameter.setRedrawParameter(redrawParameter);
+                    }
+                    final String callbackProperty = parameter.getName().replaceFirst(absolutePath, parameterRoot.getPath());
+                    final String defaultValue = parameter.getStringValue();
+                    final ActionParameter actionParameter = new ActionParameter(parameter.getName(), callbackProperty, defaultValue);
+                    action.addParameter(actionParameter);
+                });
+            }
+        });
+    }
+
+    protected abstract List<String> getRelativePaths();
     
     protected final TaCoKitElementParameter resolveParameter(final String relativePath, final Map<String, IElementParameter> settings) {
         String path = pathResolver.resolvePath(getOwnerPath(), relativePath);
