+++ main/plugins/org.talend.sdk.component.studio-integration/src/main/java/org/talend/sdk/component/studio/model/parameter/resolver/SuggestionsResolver.java
@@ -15,33 +15,25 @@
  */
 package org.talend.sdk.component.studio.model.parameter.resolver;
 
-import static java.util.Comparator.comparing;
-
-import java.util.Collection;
-import java.util.Iterator;
-import java.util.List;
-import java.util.Map;
-
-import org.talend.core.model.process.IElementParameter;
 import org.talend.sdk.component.server.front.model.ActionReference;
-import org.talend.sdk.component.server.front.model.SimplePropertyDefinition;
 import org.talend.sdk.component.studio.model.action.Action;
-import org.talend.sdk.component.studio.model.action.ActionParameter;
-import org.talend.sdk.component.studio.model.parameter.PropertyDefinitionDecorator;
 import org.talend.sdk.component.studio.model.parameter.PropertyNode;
-import org.talend.sdk.component.studio.model.parameter.TaCoKitElementParameter;
 import org.talend.sdk.component.studio.model.parameter.listener.ActionParametersUpdater;
 
-public class SuggestionsResolver extends AbstractParameterResolver {
-    
-    /**
-     * Updates action parameters whenever corresponding ElementParameters are changed
-     */
-    private final ActionParametersUpdater updater;
+import java.util.Collection;
+import java.util.List;
 
+public class SuggestionsResolver extends AbstractParameterResolver {
+
+    /**
+     * Constructor
+     *
+     * @param actionOwner
+     * @param actions
+     * @param updater ActionParameterUpdater which updates action parameters whenever corresponding ElementParameters are changed
+     */
     public SuggestionsResolver(final PropertyNode actionOwner, final Collection<ActionReference> actions, final ActionParametersUpdater updater) {
-        super(actionOwner, getActionRef(actionOwner, actions));
-        this.updater = updater;
+        super(updater.getAction(), actionOwner, getActionRef(actionOwner, actions), updater);
     }
     
     private static ActionReference getActionRef(final PropertyNode actionOwner, final Collection<ActionReference> actions) {
@@ -53,37 +45,10 @@ public class SuggestionsResolver extends AbstractParameterResolver {
                 .findFirst()
                 .orElseThrow(() -> new IllegalArgumentException("Action with name " + actionName + " wasn't found"));
     }
-    
-    /**
-     * Finds ElementParameters needed for action call by their relative path.
-     * Registers ActionParameterUpdater to each ElementParameter needed for action call
-     * Creates ActionParameter for each ElementParameter
-     * 
-     * @param settings all "leaf" Component options
-     */
-    public void resolveParameters(final Map<String, IElementParameter> settings) {
-        final Iterator<PropertyDefinitionDecorator> expectedParameters = PropertyDefinitionDecorator.wrap(actionRef.getProperties())
-                .stream()
-                .filter(p -> p.getParameter().isRoot())
-                .sorted(comparing(p -> p.getParameter().getIndex()))
-                .iterator();
-        final List<String> relativePaths = actionOwner.getProperty().getSuggestions().getParameters();
-        
-        relativePaths.forEach(relativePath -> {
-            if (expectedParameters.hasNext()) {
-                final String absolutePath = pathResolver.resolvePath(getOwnerPath(), relativePath);
-                final List<TaCoKitElementParameter> parameters = resolveParameters(absolutePath, settings);
-                final SimplePropertyDefinition parameterRoot = expectedParameters.next();
-                parameters.forEach(parameter -> {
-                    parameter.registerListener("value", updater);
-                    final String callbackProperty = parameter.getName().replaceFirst(absolutePath, parameterRoot.getPath());
-                    final String defaultValue = parameter.getStringValue();
-                    final ActionParameter actionParameter = new ActionParameter(parameter.getName(), callbackProperty, defaultValue);
-                    updater.getAction().addParameter(actionParameter);
-                });
-            }
-        });
 
+    @Override
+    protected final List<String> getRelativePaths() {
+        return actionOwner.getProperty().getSuggestions().getParameters();
     }
 
 }
